Adding test to verify loadbalancer ip of traefik svc
