+++ .gitignore
@@ -1,2 +1,3 @@
 .idea/
 .DS_Store
+.vscode/
